+++ src/02-section/table/_spans.js
@@ -3,11 +3,14 @@ const getColSpan = /colspan *?= *?["']([0-9]+)["'] *?\|?/;
 
 //colspans stretch ←left/right→
 const doColSpan = function(rows) {
-  rows.forEach((row, r) => {
+  rows.forEach((row) => {
     row.forEach((str, c) => {
       let m = str.match(getColSpan);
       if (m !== null) {
         let num = parseInt(m[1], 10);
+
+        //...maybe if num is so big, and centered, give it a new column? remove it?
+
         //splice-in n empty columns right here
         row[c] = str.replace(getColSpan, '');
         for(let i = 1; i < num; i += 1) {
@@ -28,6 +31,7 @@ const doRowSpan = function(rows) {
         let num = parseInt(m[1], 10);
         //copy this cell down n rows
         str = str.replace(getRowSpan, '');
+        row[c] = str;
         for(let i = r + 1; i < r + num; i += 1) {
           if (!rows[i]) {
             break;