+++ changelog.md
@@ -129,4 +129,4 @@
 * support way (+20%?) more templates.
 ### 7.0.0
 * change result-format in a bunch of templates
-* * notably: **reference** format, **see also**, **IPA**
+* * notably: **reference** format, **see also**, **IPA**, **main**